018-bioinformatics/requirements.txt
numpy
pytest
pytest-benchmark
python-levenshtein
